+++ gd-forum-manager-for-bbpress.php
@@ -2,14 +2,15 @@
 /**
  * Plugin Name:       GD Forum Manager: plugin for WordPress and bbPress
  * Plugin URI:        https://plugins.dev4press.com/gd-forum-manager-for-bbpress/
- * Description:       Expand how the moderators can manage forum content from the frontend, including forums and topics quick and bulk editing from any page showing list of topics or forums.
+ * Description:       Expand how the moderators can manage forum content from the frontend, from any page showing list of topics or forums.
  * Author:            Milan Petrovic
  * Author URI:        https://www.dev4press.com/
  * Text Domain:       gd-forum-manager-for-bbpress
- * Version:           2.6
+ * Version:           2.7
  * Requires at least: 5.8
- * Tested up to:      6.4
+ * Tested up to:      6.5
  * Requires PHP:      7.4
+ * Requires Plugins:  bbpress
  * License:           GPLv3 or later
  * License URI:       http://www.gnu.org/licenses/gpl-3.0.html
  *
@@ -30,26 +31,29 @@
  * along with this program. If not, see <http://www.gnu.org/licenses/>
  */
 
-use Dev4Press\v45\WordPress;
+use Dev4Press\v48\WordPress;
 
-$gdfar_dirname_basic = dirname( __FILE__ ) . '/';
-$gdfar_urlname_basic = plugins_url( '/', __FILE__ );
+const GDFAR_FILE = __FILE__;
+
+$gdfar_dirname_basic = dirname( GDFAR_FILE ) . '/';
+$gdfar_urlname_basic = plugins_url( '/', GDFAR_FILE );
 
 define( 'GDFAR_PATH', $gdfar_dirname_basic );
 define( 'GDFAR_URL', $gdfar_urlname_basic );
-define( 'GDFAR_D4PLIB', $gdfar_dirname_basic . 'd4plib/' );
+define( 'GDFAR_D4PLIB_PATH', $gdfar_dirname_basic . 'd4plib/' );
+define( 'GDFAR_D4PLIB_URL', $gdfar_urlname_basic . 'd4plib/' );
 
-require_once( GDFAR_D4PLIB . 'core.php' );
+require_once GDFAR_D4PLIB_PATH . 'core.php';
 
-require_once( GDFAR_PATH . 'core/autoload.php' );
-require_once( GDFAR_PATH . 'core/bridge.php' );
-require_once( GDFAR_PATH . 'core/functions.php' );
+require_once GDFAR_PATH . 'core/autoload.php';
+require_once GDFAR_PATH . 'core/bridge.php';
+require_once GDFAR_PATH . 'core/functions.php';
 
 gdfar_settings();
 gdfar();
 
 if ( WordPress::instance()->is_admin() ) {
-	require_once( GDFAR_PATH . 'core/admin.php' );
+	require_once GDFAR_PATH . 'core/admin.php';
 
 	gdfar_admin();
 }